requirements.txt
aiohttp
aiofiles
asyncio
dnspython
ffmpeg-python
hachoir
lyricsgenius
motor
numpy
pillow
psutil
python-dotenv
py-tgcalls==0.8.1
pyrogram==1.4.16
requests
rich
speedtest-cli
tgcrypto
yt_dlp
git+https://github.com/joetats/youtube_search@master
git+https://github.com/alexmercerind/youtube-search-python@main

